Add files via upload
The web face detector can also be designed to view faces in a video call. We need to get started with this project using an OpenCV and a webcam. OpenCV is a real-time computer vision tool. This project can be extended for use cases like user authentication in meetings, exams, police force, face unlock feature of phones, etc